Fix a Logical Error in Record Manager
There is a logical error in the class RmFileHandle. When the page was
released, the fild_hdr_ in the current RmFileHandle object will be
updated wrongly. This bug can lead to fragmentation of the storage that
is not obvious, even though the past code can pass the unit test.

This fix reveals that the unit test is not perfect as we think. The
problem of the unit test should be reported.

Signed-off-by: wold9168 <moewold@outlook.com>